imp: use memory efficient app proxies + support custom runtime methods
In addition to being more memory efficient, this allows platforms to inject custom methods into the app runtime so scripts can access new things, eg `world.platform` etc
