+++ src/core/extras/warn.js
@@ -1,0 +1,6 @@
+const warned = new Set()
+export function warn(str) {
+  if (warned.has(str)) return
+  console.warn(str)
+  warned.add(str)
+}
